+++ src/components/organisms/NewsList/NewsList.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import { FC } from 'react';
 
 import { Box } from '@mui/material';
