+++ src/components/organisms/SourceList/__tests__/SourceList.tests.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 
-import { sourceMock } from '../../../../types/__mocks__/Source.mocks';
+import { sourceMock } from '../../../../__mocks__/Source.mocks';
 import SourceList from '../SourceList';
 
 test('should render source list with just one element', () => {
